Only reap if the /tmp/clusty dir exists.

+++ web/reaper.rb
@@ -3,14 +3,16 @@ require 'date'
 require 'fileutils'
 require 'logger'
 log = Logger.new('reaper.log')
-Dir["/tmp/clusty/*"].each do |dir|
-  next unless File.directory? dir
+if Dir.exists? "/tmp/clusty"
+  Dir["/tmp/clusty/*"].each do |dir|
+    next unless File.directory? dir
 
-  stats = File::Stat.new(dir)
-  age_threshold_seconds = 60*60*24
-  age_threshold = Time.now - age_threshold_seconds
-  if stats.ctime < age_threshold
-    log.info "deleting #{dir} ctime=#{stats.ctime}"
-    FileUtils.rm_rf(dir)
+    stats = File::Stat.new(dir)
+    age_threshold_seconds = 60*60*24
+    age_threshold = Time.now - age_threshold_seconds
+    if stats.ctime < age_threshold
+      log.info "deleting #{dir} ctime=#{stats.ctime}"
+      FileUtils.rm_rf(dir)
+    end
   end
 end
